LED changes and power API

+++ include/runtime-hw/led.h
@@ -51,19 +51,20 @@ typedef enum {
 // LIFECYCLE
 
 /**
- * @brief Initialize a LED unit.
+ * @brief Initialize an LED unit.
  * @ingroup LED
  * @param gpio The GPIO pin number for the LED.
  * @param pwm Optional pointer to a PWM structure for linear control.
  * @return A LED structure representing the initialized unit.
  *
  * This function initializes an LED unit. If linear control is
- * supported, the PWM structure is valid and associated with the GPIO pin,
+ * supported (non-NULL PWM pointer associated with the GPIO pin),
  * it will set up the PWM channel for the LED.
  *
- * If the pin number is 0xFF then it assumes the on-board status LED
- * is being initialized. If the PWM structure is NULL, the LED will be
- * controlled in a simple on/off manner without PWM.
+ * If the pin number is 0xFF then it assumes the platform on-board status LED
+ * is being initialized (which may be implemented via a Wi‑Fi / radio module
+ * rather than a raw GPIO). If the PWM structure is NULL, the LED will be
+ * controlled in a simple on/off manner (binary) without PWM.
  */
 hw_led_t hw_led_init(uint8_t gpio, hw_pwm_t *pwm);
 
@@ -91,10 +92,11 @@ void hw_led_finalize(hw_led_t *led);
 // METHODS
 
 /**
- * @brief Check if an LED is valid.
+ * @brief Check if an LED handle is valid and initialized.
  * @ingroup LED
  * @param led Pointer to the LED structure to check.
- * @return True if the LED is valid, false otherwise.
+ * @return True if the LED has been successfully initialized and not yet
+ *         finalized; false otherwise.
  */
 bool hw_led_valid(hw_led_t *led);
 
@@ -102,61 +104,71 @@ bool hw_led_valid(hw_led_t *led);
  * @brief Return LED capabilities.
  * @ingroup LED
  * @param led Pointer to the LED structure to check.
- * @return The capabilities of the LED, such as whether it supports
- * linear control through PWM or binary control.
+ * @return Bitmask of @ref hw_led_cap_t values describing supported operations.
+ *
+ * Notes:
+ *  - HW_LED_CAP_BINARY implies on/off control (present for every valid LED).
+ *  - HW_LED_CAP_LINEAR implies brightness (0..255) via PWM.
+ *  - HW_LED_CAP_GPIO indicates a directly drivable GPIO (as distinct from a
+ *    virtual / module-managed status LED).
  */
 hw_led_cap_t hw_led_capabilities(hw_led_t *led);
 
 /**
- * @brief Set the LED state.
+ * @brief Set the LED binary state (full on or off).
  * @ingroup LED
  * @param led Pointer to the LED structure.
  * @param on True to turn on the LED to full brightness, false to turn it off.
- * @return True if the LED was successfully set to the desired state, false
- * otherwise.
+ * @return True on success, false on invalid handle or hardware failure.
  *
- * This function sets the state of the LED. If linear control is supported,
- * it will use the PWM channel to set the brightness level. It will cancel
- * any ongoing fade or blink operations.
+ * Cancels any ongoing blink or fade before applying the new state. For a
+ * linear (PWM) LED, "on" maps to brightness 255 and "off" maps to 0.
  */
 bool hw_led_set_state(hw_led_t *led, bool on);
 
 /**
- * @brief Set the LED brightness.
+ * @brief Set the LED brightness (linear LEDs only).
  * @ingroup LED
  * @param led Pointer to the LED structure.
- * @param brightness The brightness level to set (0-255).
+ * @param brightness Desired brightness (0 = off, 255 = full on).
+ * @return True on success (including the degenerate binary case when only
+ *         0 or 255 are meaningful), false if unsupported or invalid.
  *
- * This function sets the brightness of the LED. If linear control is supported,
- * it will use the PWM channel to set the brightness level. It will cancel
- * any ongoing fade or blink operations.
+ * Cancels any ongoing blink or fade before applying the new brightness.
+ * For non-PWM LEDs this degenerates to binary on/off; intermediate values
+ * are coerced to 0 or 255.
  */
 bool hw_led_set_brightness(hw_led_t *led, uint8_t brightness);
 
 /**
- * @brief Blink the LED continuously.
+ * @brief Blink the LED (toggle full on/off).
  * @ingroup LED
  * @param led Pointer to the LED structure.
- * @param period_ms The period of the blink in milliseconds.
- * @param repeats True to repeat the blink indefinitely, false to blink once.
- * @return True if the LED was successfully set to blink, false otherwise.
+ * @param period_ms Full ON+OFF cycle period in milliseconds.
+ * @param repeats True to continue indefinitely, false to perform a single
+ *                full cycle (ending in OFF).
+ * @return True on success, false otherwise.
  *
- * This function blinks the LED continuously with the specified period.
- * The LED will turn on and off at the specified interval.
+ *  - period_ms defines the total time for ON then OFF (two half-periods).
+ *  - The implementation enforces a minimum full cycle (currently 200 ms)
+ *    for stability; shorter requests are clamped.
+ *  - Cancels any prior blink/fade before starting.
  */
-bool hw_led_blink(hw_led_t *led, uint8_t period_ms, bool repeats);
+bool hw_led_blink(hw_led_t *led, uint32_t period_ms, bool repeats);
 
 /**
- * @brief Fade the LED continuously.
+ * @brief Fade (breathe) the LED (PWM LEDs only).
  * @ingroup LED
  * @param led Pointer to the LED structure.
- * @param period_ms The period of the fade in milliseconds.
- * @param repeats True to repeat the fade indefinitely, false to fade once.
- * @return True if the LED was successfully set to fade, false otherwise.
+ * @param period_ms Full up+down brightness cycle period in milliseconds.
+ * @param repeats True to continue indefinitely, false to perform one full
+ * cycle.
+ * @return True on success, false otherwise.
  *
- * This function fades the LED continuously with the specified period.
- * The LED will gradually change its brightness over the specified interval,
- * if it supports linear control through PWM. Otherwise, it will blink
- * instead.
+ *  - period_ms spans a full 0→255→0 traversal.
+ *  - Interval per brightness step ~ period_ms / 510 (minimum 1 ms).
+ *  - If linear (PWM) capability is absent, this function transparently
+ *    falls back to hw_led_blink() using the same period.
+ *  - Cancels any prior blink/fade before starting.
  */
-bool hw_led_fade(hw_led_t *led, uint8_t period_ms, bool repeats);
+bool hw_led_fade(hw_led_t *led, uint32_t period_ms, bool repeats);
